pyproject.toml
[tool.poetry]
name = "rec_app"
version = "2.47.a3"
description = "Recording Software for micro chips"
authors = ["Dmitri Yousef Yengej"]

[tool.poetry.dependencies]
python = "^3.10"
Kivy = "^2.1.0"
h5py = "^3.6.0"
numpy = "^1.22.3"
pycryptodome = "^3.14.1"
cryptography = "^37.0.1"
Brotli = "^1.0.9"
aiodns = "^3.0.0"
aiohttp = "^3.8.1"
psutil = "^5.9.4"
pyserial = "^3.5"
pyserial-asyncio = "^0.6"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
ipython = "^8.3.0"

[tool.poetry.group.dev.dependencies]
autopep8 = "^2.0.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
